page_table_multiarch

Struct PageTable32Cursor

pub struct PageTable32Cursor<'a, M: PagingMetaData, PTE: GenericPTE, H: PagingHandler> { /* private fields */ }

Available on 32-bit or docsrs only.

A cursor created by PageTable32::cursor to modify the page table.

Implementations

impl<'a, M: PagingMetaData, PTE: GenericPTE, H: PagingHandler> PageTable32Cursor<'a, M, PTE, H>

pub fn map( &mut self, vaddr: M::VirtAddr, target: PhysAddr, page_size: PageSize, flags: MappingFlags, ) -> PagingResult

Maps a virtual page to a physical frame with the given page_size and mapping flags.

pub fn remap( &mut self, vaddr: M::VirtAddr, paddr: PhysAddr, flags: MappingFlags, ) -> PagingResult<PageSize>

Remaps the mapping starting at vaddr, updates both the physical address and flags.

pub fn protect( &mut self, vaddr: M::VirtAddr, flags: MappingFlags, ) -> PagingResult<PageSize>

Updates the flags of the mapping starting at vaddr.

pub fn unmap( &mut self, vaddr: M::VirtAddr, ) -> PagingResult<(PhysAddr, MappingFlags, PageSize)>

Unmaps the mapping starting at vaddr.

pub fn map_region( &mut self, vaddr: M::VirtAddr, get_paddr: impl Fn(M::VirtAddr) -> PhysAddr, size: usize, flags: MappingFlags, allow_huge: bool, ) -> PagingResult

Maps a contiguous virtual memory region to a contiguous physical memory region with the given mapping flags.

pub fn unmap_region(&mut self, vaddr: M::VirtAddr, size: usize) -> PagingResult

Unmaps a contiguous virtual memory region.

pub fn protect_region( &mut self, vaddr: M::VirtAddr, size: usize, flags: MappingFlags, ) -> PagingResult

Updates mapping flags of a contiguous virtual memory region.

pub fn copy_from( &mut self, other: &PageTable32<M, PTE, H>, start: M::VirtAddr, size: usize, )

Available on crate feature copy-from only.

Copy entries from another page table within the given virtual memory range.

pub fn flush(&mut self)

Flushes the TLB according to the recorded flush requests.