page_table_multiarch
Trait PagingHandler

pub trait PagingHandler: Sized {
    // Required methods
    fn alloc_frame() -> Option<PhysAddr>;
    fn dealloc_frame(paddr: PhysAddr);
    fn phys_to_virt(paddr: PhysAddr) -> VirtAddr;
}

The low-level OS-dependent helpers that must be provided for PageTable64.

Required Methods

fn alloc_frame() -> Option<PhysAddr>

Request to allocate a 4K-sized physical frame.

fn dealloc_frame(paddr: PhysAddr)

Request to free a allocated physical frame.

fn phys_to_virt(paddr: PhysAddr) -> VirtAddr

Returns a virtual address that maps to the given physical address.

Used to access the physical memory directly in page table implementation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.