page_table_multiarch
Trait PagingMetaData

pub trait PagingMetaData: Sync + Send {
    type VirtAddr: MemoryAddr;

    const LEVELS: usize;
    const PA_MAX_BITS: usize;
    const VA_MAX_BITS: usize;
    const PA_MAX_ADDR: usize = _;

    // Required method
    fn flush_tlb(vaddr: Option<Self::VirtAddr>);

    // Provided methods
    fn paddr_is_valid(paddr: usize) -> bool { ... }
    fn vaddr_is_valid(vaddr: usize) -> bool { ... }
}

The architecture-dependent metadata that must be provided for PageTable64.

Required Associated Constants

const LEVELS: usize

The number of levels of the hardware page table.

const PA_MAX_BITS: usize

The maximum number of bits of physical address.

const VA_MAX_BITS: usize

The maximum number of bits of virtual address.

Provided Associated Constants

const PA_MAX_ADDR: usize = _

The maximum physical address.

Required Associated Types

type VirtAddr: MemoryAddr

The virtual address to be translated in this page table.

This associated type allows more flexible use of page tables structs like PageTable64, for example, to implement EPTs.

Required Methods

fn flush_tlb(vaddr: Option<Self::VirtAddr>)

Flushes the TLB.

If vaddr is None, flushes the entire TLB. Otherwise, flushes the TLB entry at the given virtual address.

Provided Methods

fn paddr_is_valid(paddr: usize) -> bool

Whether a given physical address is valid.

fn vaddr_is_valid(vaddr: usize) -> bool

Whether a given virtual address is valid.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.